enhancement + feature: program now checks that an identical, cleaned playlist hasn't already been created.
Program skips cleaning a playlist if there is already a cleaned playlist that is identical (WARNING: currently only checks to see if the first 10 songs are identical, i.e criteria for them being identical is not the entire playlist to improve runtime).

New feature that prints the average danceability/energy/loudness/speechiness/acousticness/instrumentalness of your playlist (all statistics provided by Spotify's Audio Analysis of each track, not by the program).

Issues: still resolving bugs around creating a playlist not based on a [clean] playlist, and cleaning a playlist with a near identical name.